+++ x/marker/handler.go
@@ -14,6 +14,8 @@ func NewHandler(k keeper.Keeper) sdk.Handler {
 	msgServer := keeper.NewMsgServerImpl(k)
 
 	return func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
+		ctx = ctx.WithEventManager(sdk.NewEventManager())
+
 		switch msg := msg.(type) {
 		case *types.MsgAddMarkerRequest:
 			res, err := msgServer.AddMarker(sdk.WrapSDKContext(ctx), msg)